+++ src/test/java/kitchenpos/menu/domain/MenuProductTest.java
@@ -8,8 +8,8 @@ import org.junit.jupiter.params.provider.CsvSource;
 import java.math.BigDecimal;
 
 import static kitchenpos.menu.domain.MenuTest.치킨세트;
-import static kitchenpos.domain.ProductTest.양념치킨_상품;
-import static kitchenpos.domain.ProductTest.후라이드_상품;
+import static kitchenpos.product.domain.ProductTest.양념치킨_상품;
+import static kitchenpos.product.domain.ProductTest.후라이드_상품;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class MenuProductTest {
