+++ src/test/java/kitchenpos/product/acceptance/ProductServiceTest.java
@@ -1,0 +1,4 @@
+package kitchenpos.product.acceptance;
+
+public class ProductServiceTest {
+}
